pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["felipesaadi <felipesaadisiegert@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
anthropic = "^0.39.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
